21_lesson/requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
asyncpg
Pillow
python-dotenv
httpx
pytest
pydantic-settings
aiofiles
pytest-asyncio